refactor(websocket): ws to websocket
follow the laid out file structure
move the code to the correct file
