+++ backend/internal/websocket/chat_messages.go
@@ -1,4 +1,4 @@
-package ws
+package websocket
 
 import (
 	"database/sql"
