+++ backend/internal/websocket/integration_test.go
@@ -1,4 +1,4 @@
-package ws
+package websocket
 
 import (
 	"database/sql"
