+++ backend/internal/websocket/message_persister.go
@@ -1,4 +1,4 @@
-package ws
+package websocket
 
 import (
 	"database/sql"
