+++ backend/internal/websocket/session_resolver.go
@@ -1,4 +1,4 @@
-package ws
+package websocket
 
 import (
 	"database/sql"
