+++ backend/internal/websocket/websocket_routes_test.go
@@ -1,4 +1,4 @@
-package ws
+package websocket
 
 import (
 	"bytes"
