+++ backend/internal/websocket/ws.go
@@ -1,4 +1,4 @@
-package ws
+package websocket
 
 import (
 	"encoding/json"
