+++ backend/internal/websocket/ws_test.go
@@ -1,4 +1,4 @@
-package ws
+package websocket
 
 import (
 	"testing"
